backend/requirements.txt
asgiref==3.8.1
Django==4.2.27
djangorestframework==3.15.1
djangorestframework_simplejwt==5.3.1
mysqlclient==2.1.1
pillow==9.5.0
PyJWT==2.8.0
sqlparse==0.4.4

